Factor out get-inclusion-proof command to a file
